audio: eq_iir: move iir structure definition to header file
this structure is common for both ipc3 and ipc4, move it
to header file.
